+++ src/main/java/com/m3pro/groundflip/config/RestClientConfig.java
@@ -6,8 +6,8 @@ import org.springframework.web.client.RestClient;
 
 @Configuration
 public class RestClientConfig {
-    @Bean
-    public RestClient restClient(RestClient.Builder builder) {
-        return builder.build();
-    }
+	@Bean
+	public RestClient restClient(RestClient.Builder builder) {
+		return builder.build();
+	}
 }
