+++ src/main/java/com/m3pro/groundflip/domain/dto/pixel/naverApi/Result.java
@@ -1,0 +1,15 @@
+package com.m3pro.groundflip.domain.dto.pixel.naverApi;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import lombok.Data;
+
+@Data
+@JsonIgnoreProperties(ignoreUnknown = true)
+class Result {
+
+	@JsonProperty("region")
+	private Region region;
+
+}
