+++ src/main/java/com/m3pro/groundflip/domain/entity/global/BaseTimeEntity.java
@@ -1,1 +1,1 @@
-package com.m3pro.groundflip.domain.entity.global;import java.time.LocalDateTime;import org.springframework.data.annotation.CreatedDate;import org.springframework.data.annotation.LastModifiedDate;import org.springframework.data.jpa.domain.support.AuditingEntityListener;import jakarta.persistence.Column;import jakarta.persistence.EntityListeners;import jakarta.persistence.MappedSuperclass;import lombok.Getter;@Getter@MappedSuperclass@EntityListeners(AuditingEntityListener.class)public class BaseTimeEntity {	@CreatedDate	@Column(updatable = false)	private LocalDateTime createdAt;	@LastModifiedDate	private LocalDateTime modifiedAt;}
+package com.m3pro.groundflip.domain.entity.global;import java.time.LocalDateTime;import org.springframework.data.annotation.CreatedDate;import org.springframework.data.annotation.LastModifiedDate;import org.springframework.data.jpa.domain.support.AuditingEntityListener;import jakarta.persistence.Column;import jakarta.persistence.EntityListeners;import jakarta.persistence.MappedSuperclass;import lombok.Getter;@Getter@MappedSuperclass@EntityListeners(AuditingEntityListener.class)public class BaseTimeEntity {	@CreatedDate	@Column(updatable = false)	private LocalDateTime createdAt;	@LastModifiedDate	private LocalDateTime modifiedAt;}
