+++ src/main/java/com/m3pro/groundflip/exception/AppException.java
@@ -1,1 +1,1 @@
-package com.m3pro.groundflip.exception;import lombok.AllArgsConstructor;import lombok.Getter;@Getter@AllArgsConstructorpublic class AppException extends RuntimeException {	private ErrorCode errorCode;}
+package com.m3pro.groundflip.exception;import lombok.AllArgsConstructor;import lombok.Getter;@Getter@AllArgsConstructorpublic class AppException extends RuntimeException {	private ErrorCode errorCode;}
