+++ src/main/java/com/m3pro/groundflip/repository/UserRepository.java
@@ -1,6 +1,5 @@
 package com.m3pro.groundflip.repository;
 
-import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -10,5 +9,6 @@ import com.m3pro.groundflip.enums.Provider;
 
 public interface UserRepository extends JpaRepository<User, Long> {
 	Optional<User> findByProviderAndEmail(Provider provider, String email);
+
 	Optional<User> findByNickname(String nickname);
 }
